feat: Refactor dts-process.js and add validateStationData function
