+++ lib/functions/dts-process.js
@@ -1,8 +1,7 @@
 const logger = require('../helpers/logging')
 const pg = require('../helpers/db')
 const invokeLambda = require('../helpers/invoke-lambda')
-const Joi = require('joi')
-const { deleteStation, getRloiIds, getStationData } = require('../helpers/imtd-api')
+const { deleteStation, getRloiIds, getStationData, validateStationData } = require('../helpers/imtd-api')
 
 async function insertStation (stationDataArray) {
   try {
@@ -32,23 +31,6 @@ async function getData (stationId) {
     await insertStation(stationData)
   } catch (error) {
     logger.error(`Could not process data for station ${stationId} (${error.message})`)
-  }
-}
-
-async function validateStationData (stationDataArray) {
-  const schema = Joi.object({
-    station_id: Joi.number().required(),
-    direction: Joi.string().required(),
-    display_time_series: Joi.boolean().required()
-  })
-
-  try {
-    const validatedData = await Promise.all(
-      stationDataArray.map((stationData) => schema.validateAsync(stationData))
-    )
-    return validatedData
-  } catch (error) {
-    throw new Error(`Validation error: ${error.message}`)
   }
 }
