+++ lib/helpers/imtd-api.js
@@ -3,6 +3,7 @@ const axios = require('axios')
 const { HTTP_NOT_FOUND } = require('../constants')
 const logger = require('./logging')
 const parseStation = require('../models/parse-time-series')
+const Joi = require('joi')
 
 async function deleteStation (stationId, tableName) {
   await pg(tableName).where({ station_id: stationId }).delete()
@@ -45,4 +46,21 @@ async function getStationData (stationId) {
   return []
 }
 
-module.exports = { deleteStation, getRloiIds, getImtdApiResponse, getStationData }
+async function validateStationData (stationDataArray) {
+  const schema = Joi.object({
+    station_id: Joi.number().required(),
+    direction: Joi.string().required(),
+    display_time_series: Joi.boolean().required()
+  })
+
+  try {
+    const validatedData = await Promise.all(
+      stationDataArray.map((stationData) => schema.validateAsync(stationData))
+    )
+    return validatedData
+  } catch (error) {
+    throw new Error(`Validation error: ${error.message}`)
+  }
+}
+
+module.exports = { deleteStation, getRloiIds, getImtdApiResponse, getStationData, validateStationData }
